+++ android/src/main/kotlin/net/mullvad/mullvadvpn/model/Settings.kt
@@ -6,5 +6,5 @@ data class Settings(
     var allowLan: Boolean,
     var autoConnect: Boolean,
     var tunnelOptions: TunnelOptions,
-    var showBetaReleases: Boolean?
+    var showBetaReleases: Boolean
 )